NS16550: buffer reads
This improves the performance of U-Boot when accepting rapid input,
such as pasting a sequence of commands.

Without this patch, on P4080DS I see a maximum of around 5 mw.l lines can
be pasted.  With this patch, it handles around 70 lines before lossage,
long enough for most things you'd paste.

Change-Id: Ic17bce2fe04a3526c2605ad6d58a0d1f172e728f
Signed-off-by: Scott Wood <scottwood@freescale.com>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/404
Reviewed-by: Simon Glass <sjg@chromium.org>