Drop unused file parameter when accessing nvram contents.
It seems that the nvram accessor functions have evolved and use a
parameter which is not needed anymore. This CL adds the nvram
caching in the shared memory blob and removes the unnecessary
parameter from all instances where nvram accessor functions are
used.

BUG=chromium-os:12522
TEST=manual:

. build all u-boot images, observe clean compilation

It turns out that drivers/tpm/slb9635_i2c/tpm.c:transmit_cmd() is
an unused static function, and is reported as such during
building the u-boot image. This will have to be addressed
separately.

Change-Id: Ic2a39ad151bc7801bad5b68bd44edd71d5660dcb