mmc: Tegra2: Enable dcache support by bouncing unaligned requests.
Dcache support was disabled due to dcache/dma interactions with
unaligned buffers.  When an unaligned buffer is used for DMA the
first and last few bytes of the buffer will be clobbered by the
dcache invalidate call that is required to make the contents of
the buffer visible to the CPU post DMA.  The reason that these
bytes are clobbered is that the dcache invalidate code (which is
the same as the linux kernels) checks for unaligned invalidates
and first flushes the cache lines that are being partially
invalidated.  This is required because the cache lines may be
shared with other variables, and may be dirty.  This flush however
writes over the first and last few bytes of the unaligned buffer
with whatever happened to be in the cache.

There are a number of possible solutions:

1) Modify the invalidate code to first read the partial cache line
and then invalidate and then write back just the valid part of the
line.  This suffers from a race condition with concurrent code in
interrupt handlers or other CPUs.

2) Modify all of U-Boot to allocate buffers from a block allocation
mechanism that ensures they are alligned on cache line boundaries.
While this is possible, there are some cases where the public API
passes down a buffer pointer all the way to the block read interface.
So all stand alone U-Boot programs would have to be fixed as well.

3) Use a bounce buffer that is known to be alligned.  Allocate the
buffer in the Tegra MMC code and ensure it is large enough for the
read/write as well as aligned correctly.  This solution requires an
extra memcpy of the data read or written and has a high water mark
on memory consumption.  It can be conditionally used only when the
buffer passed in is unaligned.

This patch implements the third solution.

BUG=None
TEST=Run "mmc part 0" and "ext2ls mmc 0:3" on Seaboard

Signed-off-by: Anton Staaf <robotboy@chromium.org>
Change-Id: I573ad4f0cef8888b24e1cec72a5890cd42244247
Reviewed-on: http://gerrit.chromium.org/gerrit/1836