mmc: Tegra2: Add data transfer completion timeout
Currently when no expected completion condition occures in the
mmc_send_cmd while loop that is waiting for a data transfer to
complete the MMC driver just hangs.

This patch adds an arbitrary 2 second timeout.  If nothing we
recognize occures within 2 seconds some diagnostic information
is printed and we fail out.

BUG=None
TEST=Run "mmc part 0" on Seaboard.

Signed-off-by: Anton Staaf <robotboy@chromium.org>
Change-Id: I7461040a687f4f7b477ed0bf86525fad9246ec28
Reviewed-on: http://gerrit.chromium.org/gerrit/2112