Implement recursively scanning PCI busses on the coreboot "board".
A hook is installed to configure PCI bus bridges as they encountered by u-boot.
The hook extracts the secondary bus number from the bridge's config space and
then recursively scans that bus.

BUG=chrome-os-partner:4511
TEST=Built coreboot u-boot, ran the PCI command on Alex, saw devices on bus 0,
the OXPCIe 952 on bus 1, and empty busses 2 through 5. This matches the bridges
reported on bus 0 and the PCI configuration output from coreboot.

Change-Id: I70287594aad2e3abae164db9d2293437a870c00b
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/2604
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>