CHROMIUM: [1/6] refactor for onestop; remove obsolete files
Compared to the old four firmware variants design, the onestop design is
simpler, and as a result, many parts of the old codebase are overly
complex or obsolete.

This series of commits refactor the codebase and will give us a simpler
and cleaner codebase, as our foundation of onestop firmware.

Changes in this commits:
* remove fmap.{h,c} - it was only for debugging purpose and very rarely
  useful; we should no longer maintain this feature
* remove get_firmware_body.{h,c} - since in onestop design we never have
  to load a second-stage firmware
* remove load_firmware_helper.{h,c} - same reason above

BUG=chromium-os:16508
TEST=manual

As all commits in series do not break build, you can run:

make ARCH=arm \
     CROSS_COMPILE=armv7a-cros-linux-gnueabi- \
     USE_PRIVATE_LIBGCC=yes \
     VBOOT=/build/tegra2_seaboard/usr \
     chromeos_seaboard_onestop_config

make ARCH=arm \
     CROSS_COMPILE=armv7a-cros-linux-gnueabi- \
     USE_PRIVATE_LIBGCC=yes \
     VBOOT=/build/tegra2_seaboard/usr \
     all

Change-Id: I1fcec8a2056ac8fcc63e95e35e20e20071b3d93c
Reviewed-on: http://gerrit.chromium.org/gerrit/2720
Tested-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>