Fix out of tree compilation problem.
The previously submitted fix
http://gerrit.chromium.org/gerrit/2311 introduced a new
subdirectory without taking care of the build process creating a
subdirectory for the corresponding object files.

This change modifies a makefile to have it do the right thing
(note that the changed line in the makefile was redundant before
this change, the corresponding subdirectory was created by a
Makefile one level up).

BUG=none
TEST=manual

in chroot run emerge-tegra2_kaen chromeos-u-boot-next

Observe the build pass (it was failing before).

Change-Id: I5857dc2e08777e97adcbe792e619f09263135a67
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/2756
Reviewed-by: Simon Glass <sjg@chromium.org>