Move GPIO setup to the MMC driver
At present the GPIO setup for MMC is done globally in the board file. This
will be kept in the fdt so this commit moves the GPIOs under control of
the MMC driver. This means that the initialization will only be done
if MMC is enabled, and will pick up the correct board GPIOs when required.

Note: I am not sure about this change. It is desirable to have all
device-related init in the driver with fdt. But U-Boot tends to split things
into board init (which may well init things that are never used) and the
drivers (which hope and pray that things are ready). A more invasive change
would involve packing the MMC config into a structure in the board file which
includes GPIO definitions. That seems like a bridge too far.

BUG=chromium-os:11623
TEST=build U-Boot for seaboard, mmc part 0; mmc part 1; see correct output

Change-Id: Icdfc07085bfb85be7fabf06dc662c184e77bbed9
Reviewed-on: http://gerrit.chromium.org/gerrit/2778
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>