Fix a bug determining the size of memory moved into RAM.
The code which decided what symbols to relocate based on whether or not their
targets had been relocated into RAM calculated the size of the memory moved by
subtracting two pointers. Unfortunately, these pointers were not byte sized and
the size was too small by a factor of 4. This change casts them to uintptr_t so
that the size is calculated correctly.

BUG=chrome-os-partner:3895
TEST=Built and booted with this change, and verified that odd behaviors went
away, like the "help" command working erratically.

Change-Id: Ic9e37fce73db4b896bbf8acd31a09997baa6be77
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/2855
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>