CHROMIUM: Implement firmware cache for internal context of VbExHashFirmwareBody.
fimware_cache is actually get_firmware_body_internal of the original design.
It is used as the internal context for firmware/VbExHashFirmwareBody().
It keeps the RW firmware A/B data which is loaded in VbExHashFirmwareBody().
After VbSelectFirmware() returns, we can directly jump to the loaded
firmware without another load from flash.

BUG=chromium-os:16544
TEST=build without error. pending test until vboot side is ready.

Change-Id: I25907ffa3a82505df64feb17bf8101f014ea90d4
Reviewed-on: http://gerrit.chromium.org/gerrit/3032
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>