Change how reset code is excluded.
This change reworks how the reset code is excluded so that all configuration
is done in the configuration header and not in the make file. This will make
it easier for future changes to use conditional compilation based on this
and similar options.

This version works by ifdef-ing out all of the code that would go into those
sections and all the code that refers to it. The sections are then empty, and
the linker will either leave them empty for the loader to ignore or remove
them entirely.

BUG=chrome-os-partner:3906
TEST=Built and verified that the reset code was excluded.

Change-Id: Ie9a6a49df0cea7bc2b13135dba2302368a0374c5
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/3219
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>