Force the lib_sysinfo structure to be in the .data section.
Otherwise it ends up in the .bss section. U-boot assumes that it doesn't need
to copy it over during relocation, and instead fills that whole section with
zeroes. If we really were booting from ROM that would be appropriate, but we
need some information from the coreboot tables (memory size) before then and
have to fill that structure before relocation. We skirt u-boot's assumption by
putting this in .data where it assumes there is still read only but non-zero
data.

BUG=chrome-os-partner:3906
TEST=Verified there were still coreboot values in this structure after
relocation by using the zboot command and checking the e820 map reported by
chrome os.

Change-Id: If310089a724f872920afed79a2555f079b7d4017
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/3329
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>