arm: Tegra3: clear IO_RESET bit in pinmux registers
Based on Tegra3 TRM, once E_18V bits in PMC are programmed, all IO_RESET are
needed to be cleared on LV blocks. If not, gpio settings on related lv pins
will always be set to low even if it is set to high. Specifically, it is
observed that when IO_RESET bit is not cleared in VI_D4 pinmux register,
the output of gpio on VI_D4 (PL.02) is always low. That causes LVDS shutdown
all the time.

Signed-off-by: Jimmy Zhang <jimmzhang@nvidia.com>

Change-Id: I1564af1ea5a8bf20d9f743c85015260ac342f44b
Reviewed-on: http://git-master/r/51921
Tested-by: Jimmy Zhang <jimmzhang@nvidia.com>
Reviewed-by: Tom Warren <twarren@nvidia.com>