arm: Tegra114: Dalmore: add dummy rw fdt sections
So recent changes force bundle_firmware to be called with the force_rw
flag set. This flag in turn tries to add a property to 2 nodes in the
flash map, rw-a-vblock, and rw-b-vblock. Since the space in the boot
partition on emmc on dalmore is so small, we can only fit a 1mb flashmap.

When creating this flashmap, the rw sections were removed. Now when
bundle_firmware tries to add properties, it actually forces the creation
of those nodes which are created empty except for the newly added
properties.

Then comes pack_firmware which tries to parse all nodes, finds the
2 newly created nodes, and finds they don't have 'reg' properties,
resulting in the ebuild failing.

This fix adds 2 dummy rw nodes which will pass the pack_firmware
checks.

Signed-off-by: Andrew Chew <achew@nvidia.com>
Change-Id: Ic35f3447920477729c641011ccfe1e41e30e9be5
Reviewed-on: http://git-master/r/173816
Reviewed-by: Allen Martin <amartin@nvidia.com>
(cherry picked from commit dad55509e390897bcc2eb53b795fcb940340a947)
Reviewed-on: http://git-master/r/189624
Tested-by: Jong Kim <jongk@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Jimmy Zhang <jimmzhang@nvidia.com>
Reviewed-by: Jong Kim <jongk@nvidia.com>