Tegra: Use mem size in memory controller instead of ODMData
Read MC_EMEM_CFG (offset 50h) to get EMEM_SIZE_MB, shift it
left to be ODMDATA compatible (bits 31:28), and use it instead
of ODMDATA on Venice.

ODMDATA isn't available currently because NVFlash/bootloader.bin
don't work on Venice (yet). Another option would be to modify
the CrOS script that creates the BCT and embed an ODMDATA at
that time, but this change is more forward-looking, since we
don't really use ODMDATA for anything but the RAM size bits.

Change-Id: I105f7a7d476f343b6e0a93246fb97699106a8d53
Signed-off-by: Tom Warren <twarren@nvidia.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/61511
Signed-off-by: Bryan Wu <pengw@nvidia.com>