ARM: tegra: delete /psci and /firmware nodes
The L4T DT file for Jetson TK1 contains /psci and /firmware nodes.
However, no such firmware is present in a typical L4T install. Delete
these nodes during boot so that the kernel doesn't try and use PSCI,
which would cause the system to hang or crash.

This approach is taken rather than removing the nodes from the source
DT file, to allow for the possibility of running under a secure monitor
in the future. In such a scenario, the current unconditional node
removal can be made conditional at runtime based on whether a secure
monitor is present or absent.

This approach also matches the behaviour of NVIDIA's binary bootloaders.

Change-Id: I76c3205f8d13d13b57475af19d3273f3c3ee5c77
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Reviewed-on: http://git-master/r/776411
Reviewed-by: Tom Warren <twarren@nvidia.com>