ARM: tegra: Define lp0_vec environment variable
Move the definition of the lp0_vec environment variable into a common
place, so that it applies to all chips. Actually place the value into
the environment. This will allows kernel command-lines to reference the
variable, and hence stay in sync with wherever U-Boot puts the LP0
restore code.

Bug 1652898

Change-Id: I13328e3350ba997cea7ce677f828d1654872e53d
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Reviewed-on: http://git-master/r/776415
Reviewed-by: Tom Warren <twarren@nvidia.com>