ARM: tegra: copy tboot DT to U-Boot DT
One of the primary reasons for using U-Boot is its ability to load kernel
and DTB from arbitrary locations under the control over environment
variables, scripts, or extlinux.conf menus.

However, the L4T kernel expects the DTB it receives to contain certain
runtime-generated data, such as LPDDR4 training results. These are
obviously not contained in any DTB loaded from disk or network.

Rather, tboot (which runs before U-Boot) loads a DTB from a specific
partition in the boot flash, modifies it, and passes it on to subsequent
software.

In order to pass that information to the kernel, U-Boot has two choices:

a) Pass the tboot-supplied DTB to the kernel. This had the disadvantage
that it removes all of U-Boot's flexibility in location/loading the DTB.

b) Copy the relevant information from the tboot-supplied copy of the DTB
into the U-Boot-loaded DTB before booting the kernel.

This patch implements option (b) above.

tboot stores the DTB at address 0x83080000. Hence, $ramdisk_addr_r is
modified to avoid loading any initrd over the top of the tboot-supplied
DTB.

Bug 200051179

Change-Id: I04fe1a6a64d17249a1548b13568d0122dad31609
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Reviewed-on: http://git-master/r/782016
Reviewed-by: Tom Warren <twarren@nvidia.com>