ARM: tegra: boot the BPMP firmware
The BPMP (a/k/a AVP or COP) runs a power management firmware. It is up
to U-Boot to boot that firmware. This should not happen until the last
moment before U-Boot jumps to the Linux kernel, since once the BPMP FW
is booted, its fully owns some of the SoC's I2C ports, and while we could
have U-Boot proxy some of its I2C accesses through the BPMP FW like the
Linux kernel does, it's simpler to just defer booting the BPMP.

This isn't strictly a feature related to nvtboot, but in order to know
where the firmware is, we must use information nvtboot passes us. In the
future, we might need to:
- Implement a boot_bpmp(void *fw_addr) function in a common location so
  it can be called in other cases too.
- Move board_cleanup_before_linux() into board2.c and have it call
  nvtboot_cleanup_before_linux() in addition to any other necessary
  functions.

Bug 200051179

Change-Id: If7ea690fd817a3ae8f4caf7ee2f6105a491c2ac3
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Reviewed-on: http://git-master/r/789240
GVS: Gerrit_Virtual_Submit
Reviewed-by: Tom Warren <twarren@nvidia.com>