usb: ci_udc: fix emissions of ZLPs
Commit 6a132416359e "ci_udc: Update the ci_udc driver to support bulk
transfers" caused the value of "len" to change without updating subsquent
users of that variable in ci_ep_submit_next_request(). This caused the
code that detects when to emit ZLPs (Zero Length Packets) never to
trigger, which in turn caused host timeouts when a ZLP was required,
which in turn broke tests/dfu/, even despite the assertion in that
commit's description that "These changes are tested for both the DFU and
lthor."

Fix this by modifying the added dtd iteration code not to modify "len",
but rather to keep state in a separate variable. Rename the variables
while we're at it so they describe their purpose better.

Bug 1669687

Change-Id: I81616485b8b2d36fdfe20f534b3cf0fbfe07e16c
Fixes: 6a132416359e ("ci_udc: Update the ci_udc driver to support bulk transfers")
Signed-off-by: Stephen Warren <swarren@nvidia.com>
(cherry picked from upstream mailing list)
Reviewed-on: http://git-master/r/797979
GVS: Gerrit_Virtual_Submit
Reviewed-by: Tom Warren <twarren@nvidia.com>