ARM: tegra: don't enable GPIOs until direction is set
Tegra's GPIO driver currently enables pins as GPIO as soon as they're
requested. This is not safe, since the desired direction and output value
are not yet known. This could cause a glitch on the output pins between
gpio_request() and gpio_direction_*(), depending on what values happen to
be in the GPIO controller's in/out and out-value registers vs. the final
desired configuration.

To solve this, defer enabling pins as GPIOs until some gpio_direction_*()
is invoked, and the desired configuration is explicitly programmed.

In theory this change could cause regressions, if code exists that claims
a GPIO, never explicitly sets a direction, and then gets/sets the GPIO
value based on that assumption. However, I've read through all the Tegra-
related board files and device drivers that touch GPIOs and I do not see
such buggy code anywhere.

Bug 1688711

Signed-off-by: Stephen Warren <swarren@nvidia.com>
(cherry picked from upstream mailing list)
Change-Id: Id34f2c85ad057fe548913ee5ca5c7e5f9dc26dea
Reviewed-on: http://git-master/r/803930
Reviewed-by: Tom Warren <twarren@nvidia.com>