ARM: tegra210: implement PLLE init procedure from TRM
Implement the procedure that the TRM mandates to initialize PLLREFE and
PLLE. This makes the PLL actually lock.

Note that this section of the TRM is being cleaned up to remove some
confusion. The set of register accesses in this patch should be final,
although the step numbers/descriptions might still change.

Bug 200141031

Change-Id: I42aaedb09bcee660be058d2bf4cec4d8f91aca61
Signed-off-by: Stephen Warren <swarren@nvidia.com>
(cherry-picked from upstream mailing list)
Reviewed-on: http://git-master/r/815318
GVS: Gerrit_Virtual_Submit
Reviewed-by: Tom Warren <twarren@nvidia.com>