ARM: tegra: create common XUSB padctl driver file
A fair amount of the XUSB padctl driver will be common between Tegra124
and Tegra210. To avoid cut/paste between the two chips, create a new
file that will contain the common code, and convert the Tegra124 code to
use it. This change doesn't move every last piece of code that can/will be
shared, but rather concentrates on moving code that can be moved with zero
changes, so there are no other diffs mixed in.

Bug 200141031

Change-Id: I08d0d2d2a30754c21e186324bc19ee0bedb455f4
Signed-off-by: Stephen Warren <swarren@nvidia.com>
(cherry-picked from upstream mailing list)
Reviewed-on: http://git-master/r/815320
GVS: Gerrit_Virtual_Submit
Reviewed-by: Tom Warren <twarren@nvidia.com>