ARM: tegra: prevent kernel from duplicating pinmux config
In U-Boot, for all currently supported T210 boards (and indeed some
earlier boards), U-Boot sets up the entire initial static pinmux.
However, the kernel DT contains a complete duplicate copy of the pinmux
configuration so that it can work with bootloaders that don't set up
the pinmux. When booting with U-Boot, this causes the kernel to waste
time re-applying pinmux configuration that is already programmed into
hardware. This also requires that any pinmux changes be applied in both
the U-Boot source and the kernel DT, even if it is known that U-Boot is
the only bootloader that will be used by a particular user.

We can't remove the pinmux configuration from the kernel DT, because we
still allow using a bootloader other than U-Boot.

This change enhances U-Boot's DT fixup process to delete a configurable
set of properties from DT. The default environment is set up so that
pinmux-related properties are deleted, thus preventing the kernel from
processing the base pinmux configuration. Device-specific configuration
(i.e. that referenced from any DT node other than the main pinmux node)
is still applied if present.

Bug 1738572.

Change-Id: Ibfe7bb98b228b0349b3bf4da570e1f69e2452e13
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Reviewed-on: http://git-master/r/1028191
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: Tom Warren <twarren@nvidia.com>