dm: core: Add a function to bind child devices
We currently use dm_scan_fdt_node() to bind devices. It is an internal
function and it requires the caller to know whether we are pre- or post-
relocation.

This requirement has become quite common in drivers, so the current function
is not ideal.

Add a new function with fewer arguments, that does not require internal
headers. This can be used directly as a post_bind() method if needed.

Bug 200211165

Change-Id: I5569fd4af953650c08ba326bf275eeafec1819f2
Signed-off-by: Simon Glass <sjg@chromium.org>
(cherry picked from commit cc7f66f70cc2c59fe8ebf9011658447815278894 upstream)
Reviewed-on: http://git-master/r/1195423
Tested-by: Stephen Warren <swarren@nvidia.com>
Reviewed-by: Tom Warren <twarren@nvidia.com>