misc: add "call" uclass op
The call op requests that the callee pass a message to the underlying HW
or device, wait for a response, and then pass back the response error code
and message to the callee. It is useful for drivers that represent some
kind of messaging or IPC channel to a remote device.

Bug 200211165

Change-Id: Ia2a4a860ed14e5fb0f476875560ed21cf13e3152
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Acked-by: Simon Glass <sjg@chromium.org>
(cherry picked from commit b647f55420310beb8f576e23f3b6a69745126f71 upstream)
Reviewed-on: http://git-master/r/1203150
Reviewed-by: Tom Warren <twarren@nvidia.com>
GVS: Gerrit_Virtual_Submit